Add files via upload
The Empower Rural is a visionary project leveraging web technologies to upliftrural communities by providing vital government services. It is designed using HTML and CSS for the frontend development, PHP and JavaScript for the backend development and MySQL for the database purpose. The software will be available at centers in remote areas, making essential services accessible and easy to use for the local population. Empower Rural focuses on bridging the digital divide, ensuring that even the most isolated communities can benefit from vital government services.